requirements.txt

Django
djangorestframework-simplejwt
djangorestframework
django-cors-headers
requests


